+++ cmake/modules/Element.cmake
@@ -42,6 +42,7 @@ target_sources(kwik_element
             modules/element/datepicker.cppm
             modules/element/chart.cppm
             modules/element/progressring.cppm
+            modules/element/spinbox.cppm
 
         BASE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/modules/layout
         FILES
@@ -91,6 +92,7 @@ target_sources(kwik_element
         src/element/datepicker.cpp
         src/element/chart.cpp
         src/element/progressring.cpp
+        src/element/spinbox.cpp
         
         src/layout/flex_layout.cpp
         src/layout/grid_layout.cpp